group06_version1/NCSNv2/requirements.txt
pytorch
torchvision
numpy
tqdm
scipy
matplotlib
